+++ src/App.css
@@ -16,7 +16,7 @@ body{
   background: linear-gradient(-45deg, #ee7752, #e73c7e, #23a6d5, #23d5ab);
   background-size: 400% 400%;
   animation: gradient 10s ease infinite;
-  height: 100vh;
+  min-height: 100vh;
 }  
 .App-header {
   background-color: #282c34;
